chore: split src/constants.ts into domain-specific files under src/constants/
- defaults.ts — default versions and editor view type
- patterns.ts — glob and regexp patterns
- settings.ts — general KAOTO_*_SETTING_ID constants
- executor-settings.ts — KAOTO_EXECUTOR_* setting ID constants
- commands.ts — all KAOTO_*_COMMAND_ID constants
- views.ts — view/panel container ID constants
- context-keys.ts — VS Code context key constants
- global-state-keys.ts — global state key constants
- trusted-urls.ts — trusted source URL constants
- index.ts — barrel re-exporting all of the above

All existing imports resolve via TypeScript directory→index.ts resolution.
