+++ src/constants/commands.ts
@@ -1,0 +1,93 @@
+export const COMMAND_CAMEL_NEW_FILE = 'kaoto.new.camel.file';
+
+export const COMMAND_CAMEL_ROUTE = 'kaoto.camel.jbang.init.route';
+
+export const COMMAND_CAMEL_KAMELET_YAML = 'kaoto.camel.jbang.init.kamelet.yaml';
+
+export const COMMAND_CAMEL_PIPE_YAML = 'kaoto.camel.jbang.init.pipe.yaml';
+
+export const COMMAND_CITRUS_INIT = 'kaoto.citrus.jbang.init.test';
+
+export const COMMAND_CAMEL_NEW_PROJECT = 'kaoto.camel.jbang.export';
+
+export const COMMAND_CAMEL_NEW_PROJECT_FOLDER = 'kaoto.camel.jbang.export.folder';
+
+export const COMMAND_CAMEL_NEW_PROJECT_WORKSPACE = 'kaoto.camel.jbang.export.workspace';
+
+export const COMMAND_OPENAPI_IMPORT = 'kaoto.openapi.import';
+
+export const COMMAND_UNDO = 'kaoto.undo';
+
+export const COMMAND_REDO = 'kaoto.redo';
+
+export const COMMAND_WHATS_NEW_SHOW = 'kaoto.whatsNew.show';
+
+export const COMMAND_OPEN_SOURCE = 'kaoto.open.source';
+
+export const COMMAND_CLOSE_SOURCE = 'kaoto.close.source';
+
+export const COMMAND_OPEN_WITH_KAOTO = 'kaoto.open';
+
+export const COMMAND_SELECT_CAMEL_CATALOG = 'kaoto.selectCamelCatalog';
+
+export const COMMAND_INTEGRATIONS_REFRESH = 'kaoto.integrations.refresh';
+
+export const COMMAND_INTEGRATIONS_SHOW_SOURCE = 'kaoto.integrations.showSource';
+
+export const COMMAND_INTEGRATIONS_DELETE = 'kaoto.integrations.delete';
+
+export const COMMAND_INTEGRATIONS_UPDATE_DEPENDENCIES = 'kaoto.integrations.updateDependencies';
+
+export const COMMAND_INTEGRATIONS_RUN = 'kaoto.integrations.run';
+
+export const COMMAND_INTEGRATIONS_RUN_FOLDER = 'kaoto.integrations.run.folder';
+
+export const COMMAND_INTEGRATIONS_RUN_WORKSPACE = 'kaoto.integrations.run.workspace';
+
+export const COMMAND_INTEGRATIONS_RUN_ALL_WORKSPACES = 'kaoto.integrations.run.all.workspaces';
+
+export const COMMAND_INTEGRATIONS_KUBERNETES_RUN = 'kaoto.integrations.kubernetes.run';
+
+export const COMMAND_TESTS_REFRESH = 'kaoto.tests.refresh';
+
+export const COMMAND_TESTS_CLEAR_RESULTS = 'kaoto.tests.clearResults';
+
+export const COMMAND_TESTS_RUN = 'kaoto.tests.run';
+
+export const COMMAND_TESTS_RUN_FOLDER = 'kaoto.tests.run.folder';
+
+export const COMMAND_TESTS_SHOW_SOURCE = 'kaoto.tests.showSource';
+
+export const COMMAND_TESTS_DELETE = 'kaoto.tests.delete';
+
+export const COMMAND_OPENAPI_REFRESH = 'kaoto.openapi.refresh';
+
+export const COMMAND_OPENAPI_SHOW_SOURCE = 'kaoto.openapi.showSource';
+
+export const COMMAND_OPENAPI_DELETE = 'kaoto.openapi.delete';
+
+export const COMMAND_DEPLOYMENTS_REFRESH = 'kaoto.deployments.refresh';
+
+export const COMMAND_DEPLOYMENTS_STOP = 'kaoto.deployments.stop';
+
+export const COMMAND_DEPLOYMENTS_LOGS = 'kaoto.deployments.logs';
+
+export const COMMAND_DEPLOYMENTS_ROUTE_START = 'kaoto.deployments.route.start';
+
+export const COMMAND_DEPLOYMENTS_ROUTE_STOP = 'kaoto.deployments.route.stop';
+
+export const COMMAND_DEPLOYMENTS_ROUTE_RESUME = 'kaoto.deployments.route.resume';
+
+export const COMMAND_DEPLOYMENTS_ROUTE_SUSPEND = 'kaoto.deployments.route.suspend';
+
+export const COMMAND_INFRASTRUCTURE_START = 'kaoto.infrastructure.start';
+
+export const COMMAND_INFRASTRUCTURE_STOP = 'kaoto.infrastructure.stop';
+
+export const COMMAND_INFRASTRUCTURE_REFRESH = 'kaoto.infrastructure.refresh';
+
+export const COMMAND_INFRASTRUCTURE_LOGS = 'kaoto.infrastructure.logs';
+
+export const COMMAND_INFRASTRUCTURE_COPY_URL = 'kaoto.infrastructure.copyUrl';
+
+export const COMMAND_INFRASTRUCTURE_COPY_PORT = 'kaoto.infrastructure.copyPort';
