+++ src/constants/context-keys.ts
@@ -1,0 +1,9 @@
+export const CONTEXT_EXECUTOR_AVAILABLE = 'kaoto.executorAvailable';
+
+export const CONTEXT_WORKSPACE_HAS_POM_XML = 'kaoto.workspaceHasPomXml';
+
+export const CONTEXT_INTEGRATION_EXISTS = 'kaoto.integrationExists';
+
+export const CONTEXT_TEST_EXISTS = 'kaoto.testExists';
+
+export const CONTEXT_TEST_RESULTS_EXIST = 'kaoto.testResultsExist';
