+++ src/constants/defaults.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_CAMEL_VERSION_FALLBACK: string = '4.22.0';
+
+/** Minimum camel-jbang version required for the 'camel test' plugin (4.22.0 ships the plugin natively; 4.20 uses a broken JBang shim) */
+export const MIN_CAMEL_VERSION_FOR_TEST: string = '4.22.0';
+
+export const KAOTO_EDITOR_VIEW_TYPE = 'webviewEditorsKaoto';
