+++ src/constants/executor-settings.ts
@@ -1,0 +1,21 @@
+export const KAOTO_EXECUTOR_TYPE_SETTING_ID = 'kaoto.executor.type';
+
+export const KAOTO_EXECUTOR_RUN_ARGUMENTS_SETTING_ID = 'kaoto.executor.runArguments';
+
+export const KAOTO_EXECUTOR_RUN_SOURCE_DIR_ARGUMENTS_SETTING_ID = 'kaoto.executor.runFolderOrWorkspaceArguments';
+
+export const KAOTO_EXECUTOR_RED_HAT_MAVEN_REPOSITORY_SETTING_ID = 'kaoto.executor.redHatMavenRepository';
+
+export const KAOTO_EXECUTOR_RED_HAT_MAVEN_REPOSITORY_GLOBAL_SETTING_ID = 'kaoto.executor.redHatMavenRepository.global';
+
+export const KAOTO_EXECUTOR_KUBERNETES_RUN_ARGUMENTS_SETTING_ID = 'kaoto.executor.kubernetesRunArguments';
+
+export const KAOTO_EXECUTOR_EXPORT_ARGUMENTS_SETTING_ID = 'kaoto.maven.executor.exportProjectArguments';
+
+export const KAOTO_EXECUTOR_INFRA_ARGUMENTS_SETTING_ID = 'kaoto.executor.infraArguments';
+
+export const KAOTO_RUNTIME_CATALOG_NAME_SETTING_ID = 'kaoto.runtimeCatalogName';
+
+export const KAOTO_TESTING_CATALOG_NAME_SETTING_ID = 'kaoto.testingCatalogName';
+
+export const KAOTO_INFRASTRUCTURE_ENABLED_SETTING_ID = 'kaoto.infrastructure.enabled';
