+++ src/constants/global-state-keys.ts
@@ -1,0 +1,3 @@
+export const STATE_LAST_WHATS_NEW_SHOWN_VERSION = 'kaoto.lastWhatsNewShownVersion';
+
+export const STATE_SHOW_RUN_ALL_FOLDERS_MESSAGE = 'kaoto.showRunAllFoldersMessage';
