+++ src/constants/patterns.ts
@@ -1,0 +1,5 @@
+export const KAOTO_FILE_PATH_GLOB = '**/*.{yml,yaml,xml}';
+
+export const KAOTO_EXCLUDE_PATTERN = '{**/node_modules/**,**/.vscode/**,**/out/**,**/.citrus-jbang*/**,**/.camel-jbang*/**,**/target/**,**/.mvn/**}';
+
+export const DEFAULT_KAOTO_OPENAPI_FILES_REGEXP: string[] = ['*openapi.yaml', '*openapi.yml', '*openapi.json'];
