+++ src/constants/settings.ts
@@ -1,0 +1,27 @@
+export const KAOTO_MAVEN_DEPENDENCIES_UPDATE_ON_SAVE_SETTING_ID = 'kaoto.maven.dependenciesUpdate.onSave';
+
+export const KAOTO_LOCAL_KAMELET_DIRECTORIES_SETTING_ID = 'kaoto.localKameletDirectories';
+
+export const KAOTO_INTEGRATIONS_FILES_REGEXP_SETTING_ID = 'kaoto.integrations.files.regexp';
+
+export const KAOTO_TESTS_FILES_REGEXP_SETTING_ID = 'kaoto.tests.files.regexp';
+
+export const KAOTO_OPENAPI_FILES_REGEXP_SETTING_ID = 'kaoto.openapi.files.regexp';
+
+export const KAOTO_REST_APICURIO_REGISTRY_URL_SETTING_ID = 'kaoto.restConfiguration.apicurioRegistryUrl';
+
+export const KAOTO_REST_CUSTOM_MEDIA_TYPES_SETTING_ID = 'kaoto.restConfiguration.customMediaTypes';
+
+export const KAOTO_DEPLOYMENTS_REFRESH_INTERVAL_SETTING_ID = 'kaoto.deployments.refresh.interval';
+
+export const KAOTO_VIEWS_REFRESH_INTERVAL_SETTING_ID = 'kaoto.views.refresh.interval';
+
+export const KAOTO_CATALOG_URL_SETTING_ID = 'kaoto.catalog.url';
+
+export const KAOTO_NODE_LABEL_SETTING_ID = 'kaoto.nodeLabel';
+
+export const KAOTO_NODE_TOOLBAR_TRIGGER_SETTING_ID = 'kaoto.nodeToolbarTrigger';
+
+export const KAOTO_COLOR_THEME_SETTING_ID = 'kaoto.colorTheme';
+
+export const KAOTO_CANVAS_LAYOUT_DIRECTION_SETTING_ID = 'kaoto.canvasLayoutDirection';
