+++ src/constants/trusted-urls.ts
@@ -1,0 +1,3 @@
+export const CAMEL_TRUSTED_SOURCE_URL = 'https://github.com/apache/camel/';
+
+export const CITRUS_TRUSTED_SOURCE_URL = 'https://github.com/citrusframework/citrus/';
