+++ src/constants/views.ts
@@ -1,0 +1,13 @@
+export const VIEW_HELP = 'kaoto.help';
+
+export const VIEW_INTEGRATIONS = 'kaoto.integrations';
+
+export const VIEW_TESTS = 'kaoto.tests';
+
+export const VIEW_DEPLOYMENTS = 'kaoto.deployments';
+
+export const VIEW_OPENAPI = 'kaoto.openapi';
+
+export const VIEW_INFRASTRUCTURE = 'kaoto.infrastructure';
+
+export const VIEW_WHATS_NEW = 'kaoto.whatsNew';
